Moved the call to ensureAppExists back into Server.run.
Postponed the call to reactor.suggestThreadPoolSize as late as possible,
because worker threads created by Twisted prevent the application from
aborting on exceptions in the main thread.

Discovered that unit tests in test_db leave the transaction in an unclean
state and added some transaction aborts to work around that.