Removed an offending self.groupTreeCtrl.Expand(root) line that produced assertion failures when wxPython was compiled in debug mode.
There was a slight problem with the group tree on Refresh: the selected tree
item was identified by the href of the group.  Since the same group can have
more than one parent, this ID wasn't unique and a different tree item
(representing the same group but in a different part of the tree) might get
selected after a refresh.

Solved this by assigning more complex IDs to tree items (a tuple of hrefs of
the group and all its ancestors rather than just the href of the group).
Unfortunately the code now looks hairier than I would like it to look.

The same problem exists with remembering the expanded/collapsed state of tree
items.  I added an XXX comment about that, but postponed the solution until
I think of an elegant way to express this.