Changed Template to do no encoding when charset is None and not change the Content-Type header when content_type is None.  This allows us to use Template for rendering parts that are later included in another template as parts.
Changed TreeView.node_template to specify charset and content_type to None to
make use of the above.

This fixes a possible Unicode error in TreeView where a template of the
subtree was rendered, converted to UTF-8 and then included in another
template without first converting it back to a Unicode object properly.