POST of ACLView.
Reworked the ReSTive view traversal.  Now the individual traversal names,
such as 'photo', 'calendar', 'acl', are handled by named views providing
IRestTraverser.

This commit also introduced a dependency of browser view on ReSTive views,
as the PUT requests on iCal calendars get handled by the ReSTive views.
This does not mean, however, that the ReST server has to be run.