Refactored reusable parts of the ACL benchmark into a utility module.
(Note that the code in benchmark/ has considerably lower quality standards -- it
is more important to have something that works now, than to have something that
is clear and maintainable in the long run.  Do not use it as an example of what
is acceptable for the schooltool/bell core.  I do expect that in the future we
will refactor the benchmark code somewhat if we need to write more benchmarks.)