Bugfix: quote the URL for the login form.
If you tried to access a page with a '+' in the URL (e.g. ++etc++site or an add
form), but did not have permission to do so, you'd get a login form.  If you
then logged in successfully, you'd be redirected to a nonexistent page because
all the '+' signs in the URL were replaced with spaces.

It was an old and well-known issue.  I cannot find the issue number in the
tracker -- maybe it was never filed?